+++ _build/images/base/scripts/70_install_from_pip.sh
@@ -21,6 +21,8 @@ pip3 install --upgrade --no-cache-dir --break-system-packages \
 	pathspec \
 	pipdeptree \
 	prettyprinttree \
+	prettytable \
+	pytest \
 	pyyaml \
 	scikit-build \
 	scikit-image \