+++ _build/images/base/scripts/99_basepkg_install.sh
@@ -9,17 +9,17 @@ pip3 install --upgrade --no-cache-dir \
 	jupyter-collaboration \
 	jupyterlab-night
 
+#FIXME gdfsfactory >= 8.8.6 requires Python 3.11
 echo "[INFO] Install EDA packages via PIP"
 pip3 install --upgrade --no-cache-dir \
-	amaranth==0.5.1 \
-	cace==2.3.11 \
-	cocotb==1.9.0 \
+	amaranth==0.5.3 \
+	cace==2.5.3 \
+	cocotb==1.9.2 \
 	edalize==0.5.4 \
-	fusesoc==2.3 \
-	gdsfactory==8.6.1 \
+	fusesoc==2.4 \
+	gdsfactory==8.17.0 \
 	gdspy==1.6.13 \
-	hdl21==6.0.0 \
-	openlane==2.1.0 \
+	openlane==2.2.3 \
 	openram==1.2.48 \
 	pygmid==1.2.12 \
 	pyrtl==0.11.2 \
@@ -27,27 +27,13 @@ pip3 install --upgrade --no-cache-dir \
 	pyuvm==3.0.0 \
 	pyverilog==1.3.0 \
 	schemdraw[svgmath]==0.19 \
-	scikit-rf==1.1.0 \
-	siliconcompiler==0.25.0 \
+	scikit-rf==1.3.0 \
+	siliconcompiler==0.28.4 \
 	spyci==1.0.2 \
-	vlsirtools==6.0.0 \
-	volare==0.18.1
-
-#FIXME OpenLane2 goes last to get its dependencies right
-#pip3 install --upgrade --no-cache-dir \
-#	openlane==2.0.11
-#FIXME Locally build OpenLane2 due to klayout incompatibility
-#FIXME Need to add OL2 first, otherwise dependency issue on antlr4-python3-runtime
-#cd /tmp
-#git clone --depth=1 https://github.com/iic-jku/openlane2.git
-#cd openlane2
-#pip3 install .
-#cd /
-
+	volare==0.19.2
 
 echo "[INFO] Install EDA packages via GEM"
 gem install \
 	rggen \
 	rggen-verilog \
 	rggen-vhdl
-