+++ _build/images/yosys/scripts/install.sh
@@ -8,7 +8,7 @@ git clone --filter=blob:none "${YOSYS_REPO_URL}" "${YOSYS_NAME}"
 cd "${YOSYS_NAME}" || exit 1
 git checkout "${YOSYS_REPO_COMMIT}"
 git submodule update --init
-make install -j"$(nproc)" PREFIX="${TOOLS}/${YOSYS_NAME}" CONFIG=gcc
+make install -j"$(nproc)" PREFIX="${TOOLS}/${YOSYS_NAME}" CONFIG=gcc ENABLE_PYOSYS=1
 
 export PATH=$PATH:${TOOLS}/${YOSYS_NAME}/bin
 
@@ -17,7 +17,7 @@ export PATH=$PATH:${TOOLS}/${YOSYS_NAME}/bin
 cd /tmp || exit 1
 git clone --filter=blob:none "${YOSYS_EQY_REPO_URL}" "${YOSYS_EQY_NAME}"
 cd "${YOSYS_EQY_NAME}" || exit 1
-git checkout "${YOSYS_REPO_COMMIT}"
+git checkout "yosys-${YOSYS_REPO_COMMIT}"
 sed -i "s#^PREFIX.*#PREFIX=${TOOLS}/${YOSYS_NAME}#g" Makefile
 make install -j"$(nproc)"
 
@@ -26,7 +26,7 @@ make install -j"$(nproc)"
 cd /tmp || exit 1
 git clone --filter=blob:none "${YOSYS_SBY_REPO_URL}" "${YOSYS_SBY_NAME}"
 cd "${YOSYS_SBY_NAME}" || exit 1
-git checkout "${YOSYS_REPO_COMMIT}"
+git checkout "yosys-${YOSYS_REPO_COMMIT}"
 sed -i "s#^PREFIX.*#PREFIX=${TOOLS}/${YOSYS_NAME}#g" Makefile
 make install -j"$(nproc)" 
 
@@ -35,6 +35,6 @@ make install -j"$(nproc)"
 cd /tmp || exit 1
 git clone --filter=blob:none "${YOSYS_MCY_REPO_URL}" "${YOSYS_MCY_NAME}"
 cd "${YOSYS_MCY_NAME}" || exit 1
-git checkout "${YOSYS_REPO_COMMIT}"
+git checkout "yosys-${YOSYS_REPO_COMMIT}"
 sed -i "s#^PREFIX.*#PREFIX=${TOOLS}/${YOSYS_NAME}#g" Makefile
 make install -j"$(nproc)"